# backend/src/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "API Bom click",
    "description": "API de consulta de Preços.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/BuscaUsuarios": {
      "get": {
        "summary": "Busca todos os usuários",
        "description": "Retorna a lista de usuários cadastrados no sistema.",
        "operationId": "buscaUsuarios",
        "tags": ["Usuario"],
        "responses": {
          "200": {
            "description": "Lista de usuários retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "integer", "example": 1 },
                      "nome": { "type": "string", "example": "João Silva" },
                      "email": { "type": "string", "example": "joao@email.com" }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno ao buscar usuários",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Erro ao buscar usuário."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/BuscaUsuariosUnico/{id}": {
      "post": {
        "summary": "Busca um usuário pelo ID",
        "description": "Retorna os dados de um usuário específico com base no ID fornecido.",
        "operationId": "buscaUsuarioUnico",
        "tags": ["Usuario"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID do usuário a ser buscado"
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário encontrado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string", "example": "123" },
                    "nome": { "type": "string", "example": "Maria Oliveira" },
                    "email": { "type": "string", "example": "maria@email.com" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Usuário não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Usuário não encontrado." }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno ao buscar usuário",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Erro ao buscar usuário único." }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/BuscaMercado": {
      "get": {
        "summary": "Busca  mercado",
        "description": "Retorna a lista de mercados cadastrados no sistema.",
        "operationId": "BuscaMercado",
        "tags": ["Mercado"],
        "responses": {
          "200": {
            "description": "Lista de mercado retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "integer", "example": 1 },
                      "nome": { "type": "string", "example": "Confiança" },
                      "logo": { "type": "string", "example": "https://www.confianca.com.br/logo.svg" }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno ao buscar usuários",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Erro ao buscar usuário."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}